Application development and fixes
Program name and LED configuration added to program configuration.
Program change done based on data read from FRAM.
Functions created to control LED blinking during normal operation and program creation.
Switchboard_Test function added to make default configuration externally accessible to test module.
Memory_Erase function parameter type fixed.
MEMORY_TEST changed: 'D' command added so that default configuration can be called remotely, address and bytes added to 'R' command.

+++ memory.c
@@ -152,7 +152,7 @@ bool_t Memory_Write_Data (const uint16_t address, const uint8_t data[], const ui
  * Erases n bytes starting from base address
  * Must be continuously called by the application until returns TRUE
  */
-bool_t Memory_Erase (const uint16_t address, const uint8_t bytes) {
+bool_t Memory_Erase (const uint16_t address, const uint16_t bytes) {
     bool_t status = FALSE;
 
     if (address > MEM_MAX_ADDR) {
