+++ memory.h
@@ -27,7 +27,7 @@
 void Memory_Init ();
 bool_t Memory_Read_Data (const uint16_t address, const uint8_t data[], const uint8_t bytes);
 bool_t Memory_Write_Data (const uint16_t address, const uint8_t data[], const uint8_t bytes);
-bool_t Memory_Erase (const uint16_t address, const uint8_t bytes);
+bool_t Memory_Erase (const uint16_t address, const uint16_t bytes);
 void Memory_Task ();
 bool_t Memory_Close ();
